refactor(storybook): use docs provided with component - i315
